-play, pause, stop, next, prev -play next song when media ended

+++ Player/Player.cs
@@ -16,7 +16,7 @@ namespace Player
     {
         public bool mIsPlay { get; set; }
         public int mNowPlayingOrder { get; set; }   // now playing media's Order in playlist 
-        public float mNowMediaPosition { get; set; }    // now playing media's position
+        public double mNowMediaPosition { get; set; }    // now playing media's position
         public List<MediaFile> mMediaList { get; set; }
 
         public NowPlayingList()
